#871 The partitioning follow-up: maintenance counters published, 3D and f64 tiers, Migrate's contended lines removed, the cell tree made opt-in, arrivals counted (PR #928)
Closes #910. Closes #911. Closes #912. Closes #914. Closes #926.

The follow-up to #872. Every item had to pass a measured gate; what failed its gate was removed rather than shipped.

#911 publishes the spatial maintenance counters that were computed and thrown away, splits the fence into spans a trace can attribute, and adds three ClusterMigration profiler gates with their regenerated telemetry reference.

#912 and #926 remove two cache lines that every worker took exclusive per migrated entity. The source cluster's flag writes move to the per-chunk drain (−38 % Migrate CPU at W = 8), and the zone map's grow latch is taken once per slice instead of once per migrant (−64 to −76 % at W = 8). Migrate at W = 8 now costs 1.27× W = 2 per entity.

#914 and #919 make WriteSpatial handle all eight spatial tiers, with the real Z centre, Z shrink bits and a reachable Z outlier guard; the f64 tiers write, configure and query. A 3D archetype on the write barrier halves Prep against the MVCC path and writes no WAL frame per entity.

#916 pools the query enumerator's 1 KB traversal stack, rented on first descent, and deletes a descriptor copy nothing read.

#917 and #918 measured step 17. D7 split/merge and the self-measuring promotion gate were built, measured and removed, and D6 was not built. The per-cell R-Tree stays as an opt-in; CellTreePromoteThreshold now defaults to never.

#906 adds a per-system MinChunkSize for systems whose per-entity cost is far above the schedule's average.

#910 ships T0: jump and clamp counters, the largest arrival run, and a rate-limited out-of-world warning. The drain sort moves into the Prep tail and covers only the drain prefix, so the serial fence now places in the parallel fence's order. T1 and T2 are not built: gate G1 failed on SWG Tatooine's shuttles, and the synthetic worst case only matters at about 2 000 arrivals into one cell.

Two test fixes ride along: the subscription tests no longer take their ports from a constant, and the B+Tree stress dumps format numbers independently of the developer's locale.

+++ demo/AntHill/AntHill.Core/Engine/TyphonBridge.cs
@@ -2096,16 +2096,16 @@ public sealed class TyphonBridge : IDisposable
                                     // Phase 6 polish — Fight alarm phero (3×3 stamp) at the victim
                                     // position so the colony "hears the scream". Drives workers to flee
                                     // and soldiers to converge on the spider's neighbourhood.
-                                    PheromoneGrid.DepositArea(_pheromones.Fight, hit.MinX, hit.MinY, FightDeposit);
+                                    PheromoneGrid.DepositArea(_pheromones.Fight, (float)hit.MinX, (float)hit.MinY, FightDeposit);
                                     killsLeft--;
                                     killsDone++;
                                     continue;
                                 }
                                 if (d2 < bestD2)
                                 {
-                                    bestD2 = d2;
-                                    bestX = ax;
-                                    bestY = ay;
+                                    bestD2 = (float)d2;
+                                    bestX = (float)ax;
+                                    bestY = (float)ay;
                                     foundTarget = true;
                                 }
                             }
